pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.scripts]
ansifier = "ansifier.cli:main"

[project]
name = "ansifier"
description = "A python package used to represent static & animated images as text"
version = "0.1.0"
dependencies = [
  "Pillow>=9.0.1",
  "colorama==0.4.6"
]
requires-python = ">=3.8"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Natural Language :: English",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Artistic Software",
    "Topic :: Utilities"
] 
authors = [
  { name="Amelia Miner", email="ameliamminer@gmail.com" },
]

[project.urls]
Homepage = "https://github.com/amminer/ansifier"
Issues = "https://github.com/amminer/ansifier/issues"

